move EthLeverageWallet to src folder, add __init__.py and update test.py accordingly
